requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.2.1=h06a4308_0
certifi=2021.10.8=py37h06a4308_2
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.10.0=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fsspec=2022.2.0=pyhd3eb1b0_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9b=h024ee3a_2
kiwisolver=1.3.2=py37h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.2=h7f8727e_0
locket=0.2.1=py37h06a4308_2
lz4-c=1.9.3=h295c915_1
matplotlib-base=3.5.1=py37ha18d171_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd3eb1b0_0
ninja=1.10.2=py37hd09550d_3
numpy-base=1.21.2=py37h79a1101_0
olefile=0.46=py37_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_1
pillow=8.0.0=py37h9a89aac_0
pip=21.2.2=py37h06a4308_0
pyparsing=3.0.4=pyhd3eb1b0_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.8.0=py3.7_cuda10.2_cudnn7.6.5_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=6.0=py37h7f8727e_1
readline=8.1.2=h7f8727e_1
scikit-image=0.15.0=py37hb3f55d8_2
scipy=1.7.3=py37hc147768_0
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.0=hc218d9a_0
tk=8.6.11=h1ccaba5_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.8.0=py37
tqdm=4.63.0=pyhd8ed1ab_0
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0

pillow~=12.0.0
scikit-image~=0.26.0
tqdm~=4.67.1
opencv-python~=4.12.0.88

fastapi
[standard]
python-multipart
numpy
torch
torchvision
huggingface_hub